requirements.txt
# ── Core ──────────────────────────────────────────────────────────
crewai==1.14.1
crewai-tools==1.14.1

# ── LLM Providers ─────────────────────────────────────────────────
langchain==0.3.0
langchain-core==0.3.0
langchain-ollama
langchain-openai
langchain-anthropic

# ── Utilities ─────────────────────────────────────────────────────
python-dotenv
pydantic>=2.0.0
